Moved components into their own files
